Split mutation
