+++ apps/meteor/client/components/message/toolbar/useStarMessageAction.tsx
@@ -1,20 +1,15 @@
 import type { IRoom } from '@rocket.chat/core-typings';
 import { isOmnichannelRoom } from '@rocket.chat/core-typings';
-import { useSetting, useToastMessageDispatch } from '@rocket.chat/ui-contexts';
-import { useQueryClient } from '@tanstack/react-query';
+import { useSetting } from '@rocket.chat/ui-contexts';
 import { useEffect } from 'react';
 
 import { MessageAction } from '../../../../app/ui-utils/client/lib/MessageAction';
-import { sdk } from '../../../../app/utils/client/lib/SDKClient';
-import { toggleStarredMessage } from '../../../lib/mutationEffects/starredMessage';
-import { roomsQueryKeys } from '../../../lib/queryKeys';
+import { useStarMessageMutation } from '../hooks/useStarMessageMutation';
 
 export const useStarMessageAction = (room: IRoom) => {
-	const dispatchToastMessage = useToastMessageDispatch();
-
 	const allowStarring = useSetting('Message_AllowStarring');
 
-	const queryClient = useQueryClient();
+	const { mutateAsync: starMessage } = useStarMessageMutation();
 
 	useEffect(() => {
 		if (!allowStarring || isOmnichannelRoom(room)) {
@@ -30,15 +25,7 @@ export const useStarMessageAction = (room: IRoom) => {
 			type: 'interaction',
 			context: ['starred', 'message', 'message-mobile', 'threads', 'federated', 'videoconf', 'videoconf-threads'],
 			async action(_, { message }) {
-				try {
-					await sdk.rest.post('/v1/chat.starMessage', { messageId: message._id });
-					toggleStarredMessage(message, true);
-				} catch (error) {
-					dispatchToastMessage({ type: 'error', message: error });
-				} finally {
-					queryClient.invalidateQueries(roomsQueryKeys.starredMessages(message.rid));
-					queryClient.invalidateQueries(roomsQueryKeys.messageActions(message.rid, message._id));
-				}
+				await starMessage(message);
 			},
 			condition({ message, user }) {
 				return !Array.isArray(message.starred) || !message.starred.find((star: any) => star._id === user?._id);
@@ -50,5 +37,5 @@ export const useStarMessageAction = (room: IRoom) => {
 		return () => {
 			MessageAction.removeButton('star-message');
 		};
-	}, [allowStarring, dispatchToastMessage, queryClient, room]);
+	}, [allowStarring, room, starMessage]);
 };
